% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb_query.R
\name{hdb_query}
\alias{hdb_query}
\title{Query HDB}
\usage{
hdb_query(
  sdi,
  server = c("lc", "uc", "ecao", "yao", "lbao"),
  time_step = c("instantaneous", "hourly", "daily", "monthly"),
  start_date,
  end_date,
  mrid = NULL
)
}
\arguments{
\item{sdi}{Vector of all sdis to query from the specified server. It is up
to the user to know what sdis are required and what server they are found
on.}

\item{server}{The (single) server to use.}

\item{time_step}{The (single) time step of the data. Can also use the first
letter, i.e., "i", "h", "d", or "m", respectively.}

\item{start_date, end_date}{Start and end dates as character scalars. Must be
specified to the highest precision necessary for the \code{time_step} in
"yyyy-mm-dd hh:mm" format. See \strong{Details}.}

\item{mrid}{The model run ID. \code{NULL} to query observed data, otherwise
specify the model run ID as a single numerical value.}
}
\description{
Query Bureau of Reclamation's public hydrologic database (HDB). A UI is
available at https://www.usbr.gov/lc/region/g4000/riverops/_HdbWebQuery.html.
Both observed data and modeled data can be queried.
}
\details{
For specifying the start and end dates, they must be specified at the highest
precision required for the \code{time_step}, i.e., year and month for monthly,
year, month, day for daily, and year, month, day, hour minute for hourly and
instantaneous. Should be specified in yyyy-mm-dd hh:mm format. Days are
ignored if specified for monthly data.
}
