% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhdb.R
\docType{package}
\name{rhdb-package}
\alias{rhdb}
\alias{rhdb-package}
\title{rhdb: Query Data from Bureau of Reclamation's Hydrologic Database
        (HDB)}
\description{
R interface to query Bureau of Reclamation's public facing
        Hydrologic Database (HDB), available at:
        <https://www.usbr.gov/lc/region/g4000/riverops/_HdbWebQuery.html/>.
}
\details{
rhdb includes the following package options:
\itemize{
\item rhdb.try_url_n - this specifies the number of times the API url will be
tried before aborting with an error. The default is 5.
}
}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov}

}
\keyword{internal}
