% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdb_metadata.R
\name{hdb_metadata}
\alias{hdb_metadata}
\title{Get hdb meta data}
\usage{
hdb_metadata(server = "uc", type = "reservoir")
}
\arguments{
\item{server}{Character scalar of server. Only "uc" or "lc" is valid.}

\item{type}{Character scalar of data type. Only "reservoir" is valid.}
}
\value{
data frame
}
\description{
\code{hdb_metadata()} gets the metadata for a particular hdb server and data type
(e.g., reservoir). This allows the user to find the correct SDI when knowing
a site name and variable type, rather than having to know the SDI directly.
\strong{Currently, this only works for UC and LC reservoir data.} The data frame
has different variables depending on whether it is UC or LC.
}
